// src/core/server/utility/enums.ts
export enum TlrpEvent {
    VEHICLE_DESPAWNED = 'tlrp:VehicleDespawned',
    VEHICLE_ENGINE_STATE_CHANGE = 'tlrp:VehicleEngineState',
    VEHICLE_LOCK_STATE_CHANGE = 'tlrp:VehicleLockState',
    VEHICLE_SPAWNED = 'tlrp:VehicleSpawned',
    VEHICLE_REPAIRED = 'tlrp:VehicleRepaired',
    PLAYER_UNCONSCIOUSE = 'tlrp:PlayerUnconsciouse',
    PLAYER_DROPPED_ITEM = 'tlrp:PlayerDroppedItem',
    PLAYER_SELECTED_CHARACTER = 'tlrp:SelectedCharacter',
    PLAYER_SPAWNED = 'tlrp:PlayerSpawned',
    TLRP_READY = 'tlrp:Ready'
}